requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-adminlte3==0.1.6
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-html==0.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.4.2
filters==1.3.2
frozenlist==1.2.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==5.2.0
openapi-codec==1.3.2
Pillow==8.4.0
platformdirs==2.4.1
psycopg2==2.9.2
PyJWT==2.3.0
pyTelegramBotAPI==4.5.1
python-dateutil==1.5
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
telegram==0.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.0
whitenoise==6.0.0
yarl==1.7.2
